+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/ChatDomain.kt
@@ -62,7 +62,7 @@ public sealed interface ChatDomain {
     public val connectionState: StateFlow<ConnectionState>
 
     /**
-     * StateFlow<Boolean> that indicates if we are currently online, connecting of offline.
+     * StateFlow<Boolean> that indicates if we are currently online, connecting or offline.
      */
     @Deprecated("Use connectionState instead")
     public val online: StateFlow<Boolean>