+++ CHANGELOG.md
@@ -93,12 +93,18 @@ Consider migrating to `stream-chat-android-ui-components` or `stream-chat-androi
 - Fixed the information about channel members shown in the `MessageListHeader` subtitle.
 
 ### ⬆️ Improved
+- Updated a lot of documentation around the Messages features
 
 ### ✅ Added
 - Added the "mute" option to the `ChannelInfo` action dialog.
+- Added a wrapper for the message input state in the form of `MessageInputState`
 - Added `attachmentsContentImageWidth`, `attachmentsContentImageHeight`, `attachmentsContentGiphyWidth`, `attachmentsContentGiphyHeight`, `attachmentsContentLinkWidth`, `attachmentsContentFileWidth` and `attachmentsContentFileUploadWidth` options to `StreamDimens`, to make it possible to customize the dimensions of attachments content via `ChatTheme`.
 
 ### ⚠️ Changed
+- Made the MessageMode subtypes to the parent class, to make it easier to understand when importing
+- Renamed the MessageMode.Thread to MessageMode.MessageThread for clarity
+- Changed the signature of the MessageComposer to accommodate for the `MessageInputState`
+- Moved common state to the `io.getstream.chat.android.common` package
 - Made the `AttachmentFactory.previewContent` field nullable.
 
 ### ❌ Removed